+++ .setup/cpp/cTMed-total-std-vec.cpp
@@ -22,20 +22,9 @@ arma::vec TotalStdVec(const arma::vec& v, const double& delta_t) {
       index++;
     }
   }
-  arma::mat I = arma::eye(p, p);
-  arma::mat J = arma::eye(p * p, p * p);
   arma::mat total = arma::expmat(delta_t * phi);
-  arma::mat phi_hashtag = arma::kron(phi, I) + arma::kron(I, phi);
-  arma::vec sigma_vec = arma::vectorise(sigma);
-  // arma::vec psi_vec = arma::inv(phi_hashtag) * (arma::expmat(phi_hashtag *
-  // delta_t) - J) * sigma_vec;
-  arma::vec psi_vec = arma::solve(
-      phi_hashtag, (arma::expmat(phi_hashtag * delta_t) - J) * sigma_vec);
-  // arma::mat total_cov = arma::reshape(arma::inv(J - arma::kron(total, total))
-  // * psi_vec, phi.n_rows, phi.n_cols);
-  arma::mat total_cov =
-      arma::reshape(arma::solve(J - arma::kron(total, total), psi_vec),
-                    phi.n_rows, phi.n_cols);
+  arma::mat total_cov;
+  arma::syl(total_cov, phi, phi.t(), sigma * sigma.t());
   arma::mat sd_row = arma::diagmat(arma::sqrt(total_cov.diag()));
   arma::mat sd_col_inv = arma::diagmat(1.0 / arma::sqrt(total_cov.diag()));
   arma::mat total_std = sd_row * total * sd_col_inv;